% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmx.R
\name{cmx}
\alias{cmx}
\title{Title Correlation Lower Matrix}
\usage{
cmx(x)
}
\arguments{
\item{x}{Numeric values required}
}
\value{
Output displays Lower matrix of Correlation
}
\description{
Title Correlation Lower Matrix
}
