% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Npsem.R
\name{lcm_Npsem}
\alias{lcm_Npsem}
\title{R6 class for an lcm_Npsem}
\description{
R6 class for an lcm_Npsem

R6 class for an lcm_Npsem
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{lcm_Npsem$new()}}
\item \href{#method-history}{\code{lcm_Npsem$history()}}
\item \href{#method-all_vars}{\code{lcm_Npsem$all_vars()}}
\item \href{#method-clone}{\code{lcm_Npsem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lcm_Npsem$new(W, L, A, Z, M, Y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-history"></a>}}
\if{latex}{\out{\hypertarget{method-history}{}}}
\subsection{Method \code{history()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lcm_Npsem$history(var = c("L", "A", "Z", "M", "Y"), t)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_vars"></a>}}
\if{latex}{\out{\hypertarget{method-all_vars}{}}}
\subsection{Method \code{all_vars()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lcm_Npsem$all_vars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{lcm_Npsem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
