% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossFit.R
\name{Regress}
\alias{Regress}
\title{Train an sl3 object}
\usage{
Regress(Tr, y, x, type, lrnrs)
}
\arguments{
\item{Tr}{training data}

\item{y}{name of outcome}

\item{x}{names of covariates}

\item{type}{outcome type}

\item{lrnrs}{sl3 learners}
}
\description{
Train an sl3 object
}
